% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_list.R
\name{apa_add_polygon}
\alias{apa_add_polygon}
\title{Add \code{POLYGON}-columns to an \code{apa_list}}
\usage{
apa_add_polygon(apa_list)
}
\arguments{
\item{apa_list}{A \code{apa_list}-object that was created with the
[apa_list]-function.}
}
\description{
Add \code{POLYGON}-columns to an \code{apa_list}
}
