% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_length.R
\name{boundary_length}
\alias{boundary_length}
\title{Calculate boundary lengths between classes in Raster*-objects}
\usage{
boundary_length(
  rst,
  layer = NULL,
  subplot = NULL,
  plot_id_column = NULL,
  critical_layer = NULL,
  radius = 10,
  remove_na = FALSE,
  drop_geometries = TRUE,
  border_tree = FALSE
)
}
\description{
Takes a single or a list of Raster*-objects as input and returns border
lengths between classes.
}
\keyword{internal}
