% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plot_enrico}
\alias{plot_enrico}
\title{Plot-level data of 16 forest stands in Lower Saxony, Germany}
\format{
A \code{data.frame} of 16 plots. Plot coordinates are stored as
'simple feature column' (see the \code{\link[sf:sf]{sf}}-package for more details):

\describe{

\item{id_plot}{Unique plot id's.}

\item{border_geometry}{A \code{sfc_POLYGON}-column specifying the
coordinates of the plot border relative to the plot center.}

\item{buffer_geometry}{A \code{sfc_POLYGON}-column specifying the
coordinates of the plot buffer (10 m distance from \code{boder_geometry})
relative to the plot center.}
}
}
\usage{
plot_enrico
}
\description{
This dataset contains plot-level data from the Enrico-project from 16 mixed
forest stands in Lower Saxony, Germany.
}
\references{
{ Ammer, Christian; Annighoefer, Peter; Balkenhol, Niko; Hertel,
 Dietrich; Leuschner, Christoph; Polle, Andrea; Lamersdorf, Norbert; Scheu,
 Stefan; Glatthorn, Jonas (2020): RTG 2300 - Enrichment of European beech
 forests with conifers. PANGAEA, https://doi.org/10.1594/PANGAEA.925228}
}
\keyword{datasets}
