% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_polygon.R
\name{rasterize_polygon}
\alias{rasterize_polygon}
\title{Rasterize one or multiple polygons. Takes sf or sp polygons as inputs and
returns a list with one \code{RasterLayer} per polygon.}
\usage{
rasterize_polygon(polygon, res = 1, labels = "layer")
}
\description{
Rasterize one or multiple polygons. Takes sf or sp polygons as inputs and
returns a list with one \code{RasterLayer} per polygon.
}
\keyword{internal}
