% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subplot_enrico}
\alias{subplot_enrico}
\title{Subplot-level data of 16 forest stands in Lower Saxony, Germany}
\format{
A \code{data.frame} of 352 subplots on 16 plots. Plot coordinates are
stored as 'simple feature column' (see the \code{\link[sf:sf]{sf}}-package for more
details):

\describe{

\item{id_subplot}{Unique subplot id's.}

\item{id_plot}{Plot id's.}

\item{object}{Name of the subplots ('plot', 'buffer', 'quadrant',
'sixteenth').}

\item{geometry}{A \code{sfc_POLYGON}-column specifying the coordinates of
the subplots relative to the plot center.}

}
}
\usage{
subplot_enrico
}
\description{
This dataset contains subplot-level data from the Enrico-project from 16
mixed forest stands in Lower Saxony, Germany.
}
\references{
{ Ammer, Christian; Annighoefer, Peter; Balkenhol, Niko; Hertel,
 Dietrich; Leuschner, Christoph; Polle, Andrea; Lamersdorf, Norbert; Scheu,
 Stefan; Glatthorn, Jonas (2020): RTG 2300 - Enrichment of European beech
 forests with conifers. PANGAEA, https://doi.org/10.1594/PANGAEA.925228}
}
\keyword{datasets}
