% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_list.R
\name{subset.apa_list}
\alias{subset.apa_list}
\title{Subsetting of apa_lists}
\usage{
\method{subset}{apa_list}(x, subset, ...)
}
\arguments{
\item{x}{A /code{apa_list}-object.}

\item{subset}{Either a numeric or a character vector specifying the plots that
shall be selected.}

\item{...}{not implemented.}
}
\description{
Select one or more APA-maps and related data from an \code{apa_list} object.
}
